chore: refactor dependency injection
